+++ examples/example_9_wide_resource_spacing_multiple/ex_9_model_classes.py
@@ -175,7 +175,7 @@ class Model:
             self.logger.log_resource_use_start(
                 entity_id=patient.id,
                 event=f"{ward}_stay_begins",
-                resource_id=bed_resource.id_attribute
+                resource_id=bed_resource.id
                 )
 
             # sample treatment duration
@@ -184,7 +184,7 @@ class Model:
             self.logger.log_resource_use_end(
                 entity_id=patient.id,
                 event=f"{ward}_stay_complete",
-                resource_id=bed_resource.id_attribute
+                resource_id=bed_resource.id
                 )
 
         self.logger.log_departure(